+++ mmdet/ops/dcn/__init__.py
@@ -1,10 +1,9 @@
-from .functions import deform_conv
-from .functions.modulated_dcn_func import (modulated_deform_conv,
-                                           deform_roi_pooling)
-from .modules.deform_conv import DeformConv
-from .modules.modulated_dcn import (
-    DeformRoIPooling, ModulatedDeformRoIPoolingPack, ModulatedDeformConv,
-    ModulatedDeformConvPack)
+from .functions.deform_conv import deform_conv, modulated_deform_conv
+from .functions.deform_pool import deform_roi_pooling
+from .modules.deform_conv import (DeformConv, ModulatedDeformConv,
+                                  ModulatedDeformConvPack)
+from .modules.deform_pool import (DeformRoIPooling,
+                                  ModulatedDeformRoIPoolingPack)
 
 __all__ = [
     'DeformConv', 'DeformRoIPooling', 'ModulatedDeformRoIPoolingPack',
