+++ mmdet/ops/dcn/functions/deform_conv.py
@@ -3,6 +3,7 @@ from torch.autograd import Function
 from torch.nn.modules.utils import _pair
 
 from .. import deform_conv_cuda
+from .. import modulated_dcn_cuda as _backend
 
 
 class DeformConvFunction(Function):
@@ -37,13 +38,6 @@ class DeformConvFunction(Function):
         if not input.is_cuda:
             raise NotImplementedError
         else:
-            if isinstance(input, torch.autograd.Variable):
-                if not isinstance(input.data, torch.cuda.FloatTensor):
-                    raise NotImplementedError
-            else:
-                if not isinstance(input, torch.cuda.FloatTensor):
-                    raise NotImplementedError
-
             cur_im2col_step = min(ctx.im2col_step, input.shape[0])
             assert (input.shape[0] %
                     cur_im2col_step) == 0, 'im2col step must divide batchsize'
@@ -63,13 +57,6 @@ class DeformConvFunction(Function):
         if not grad_output.is_cuda:
             raise NotImplementedError
         else:
-            if isinstance(grad_output, torch.autograd.Variable):
-                if not isinstance(grad_output.data, torch.cuda.FloatTensor):
-                    raise NotImplementedError
-            else:
-                if not isinstance(grad_output, torch.cuda.FloatTensor):
-                    raise NotImplementedError
-
             cur_im2col_step = min(ctx.im2col_step, input.shape[0])
             assert (input.shape[0] %
                     cur_im2col_step) == 0, 'im2col step must divide batchsize'
@@ -112,4 +99,70 @@ class DeformConvFunction(Function):
         return output_size
 
 
+class ModulatedDeformConvFunction(Function):
+
+    @staticmethod
+    def forward(ctx,
+                input,
+                offset,
+                mask,
+                weight,
+                bias,
+                stride,
+                padding,
+                dilation=1,
+                deformable_groups=1):
+        ctx.stride = stride
+        ctx.padding = padding
+        ctx.dilation = dilation
+        ctx.deformable_groups = deformable_groups
+        if not input.is_cuda:
+            raise NotImplementedError
+        if weight.requires_grad or mask.requires_grad or offset.requires_grad \
+                or input.requires_grad:
+            ctx.save_for_backward(input, offset, mask, weight, bias)
+        output = input.new(
+            *ModulatedDeformConvFunction._infer_shape(ctx, input, weight))
+        ctx._bufs = [input.new(), input.new()]
+        _backend.modulated_deform_conv_cuda_forward(
+            input, weight, bias, ctx._bufs[0], offset, mask, output,
+            ctx._bufs[1], weight.shape[2], weight.shape[3], ctx.stride,
+            ctx.stride, ctx.padding, ctx.padding, ctx.dilation, ctx.dilation,
+            ctx.deformable_groups)
+        return output
+
+    @staticmethod
+    def backward(ctx, grad_output):
+        if not grad_output.is_cuda:
+            raise NotImplementedError
+        input, offset, mask, weight, bias = ctx.saved_tensors
+        grad_input = torch.zeros_like(input)
+        grad_offset = torch.zeros_like(offset)
+        grad_mask = torch.zeros_like(mask)
+        grad_weight = torch.zeros_like(weight)
+        grad_bias = torch.zeros_like(bias)
+        _backend.modulated_deform_conv_cuda_backward(
+            input, weight, bias, ctx._bufs[0], offset, mask, ctx._bufs[1],
+            grad_input, grad_weight, grad_bias, grad_offset, grad_mask,
+            grad_output, weight.shape[2], weight.shape[3], ctx.stride,
+            ctx.stride, ctx.padding, ctx.padding, ctx.dilation, ctx.dilation,
+            ctx.deformable_groups)
+
+        return (grad_input, grad_offset, grad_mask, grad_weight, grad_bias,
+                None, None, None, None)
+
+    @staticmethod
+    def _infer_shape(ctx, input, weight):
+        n = input.size(0)
+        channels_out = weight.size(0)
+        height, width = input.shape[2:4]
+        kernel_h, kernel_w = weight.shape[2:4]
+        height_out = (height + 2 * ctx.padding -
+                      (ctx.dilation * (kernel_h - 1) + 1)) // ctx.stride + 1
+        width_out = (width + 2 * ctx.padding -
+                     (ctx.dilation * (kernel_w - 1) + 1)) // ctx.stride + 1
+        return n, channels_out, height_out, width_out
+
+
 deform_conv = DeformConvFunction.apply
+modulated_deform_conv = ModulatedDeformConvFunction.apply
