+++ mmdet/ops/dcn/modules/deform_pool.py
@@ -1,97 +1,6 @@
-#!/usr/bin/env python
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-import math
-
-import torch
 from torch import nn
-from torch.nn.modules.utils import _pair
 
-from ..functions.modulated_dcn_func import deform_roi_pooling
-from ..functions.modulated_dcn_func import modulated_deform_conv
-
-
-class ModulatedDeformConv(nn.Module):
-
-    def __init__(self,
-                 in_channels,
-                 out_channels,
-                 kernel_size,
-                 stride,
-                 padding,
-                 dilation=1,
-                 deformable_groups=1,
-                 no_bias=True):
-        super(ModulatedDeformConv, self).__init__()
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-        self.kernel_size = _pair(kernel_size)
-        self.stride = stride
-        self.padding = padding
-        self.dilation = dilation
-        self.deformable_groups = deformable_groups
-        self.no_bias = no_bias
-
-        self.weight = nn.Parameter(
-            torch.Tensor(out_channels, in_channels, *self.kernel_size))
-        self.bias = nn.Parameter(torch.zeros(out_channels))
-        self.reset_parameters()
-        if self.no_bias:
-            self.bias.requires_grad = False
-
-    def reset_parameters(self):
-        n = self.in_channels
-        for k in self.kernel_size:
-            n *= k
-        stdv = 1. / math.sqrt(n)
-        self.weight.data.uniform_(-stdv, stdv)
-        self.bias.data.zero_()
-
-    def forward(self, input, offset, mask):
-        return modulated_deform_conv(input, offset, mask, self.weight,
-                                     self.bias, self.stride, self.padding,
-                                     self.dilation, self.deformable_groups)
-
-
-class ModulatedDeformConvPack(ModulatedDeformConv):
-
-    def __init__(self,
-                 in_channels,
-                 out_channels,
-                 kernel_size,
-                 stride,
-                 padding,
-                 dilation=1,
-                 deformable_groups=1,
-                 no_bias=False):
-        super(ModulatedDeformConvPack,
-              self).__init__(in_channels, out_channels, kernel_size, stride,
-                             padding, dilation, deformable_groups, no_bias)
-
-        self.conv_offset_mask = nn.Conv2d(
-            self.in_channels,
-            self.deformable_groups * 3 * self.kernel_size[0] *
-            self.kernel_size[1],
-            kernel_size=self.kernel_size,
-            stride=(self.stride, self.stride),
-            padding=(self.padding, self.padding),
-            bias=True)
-        self.init_offset()
-
-    def init_offset(self):
-        self.conv_offset_mask.weight.data.zero_()
-        self.conv_offset_mask.bias.data.zero_()
-
-    def forward(self, input):
-        out = self.conv_offset_mask(input)
-        o1, o2, mask = torch.chunk(out, 3, dim=1)
-        offset = torch.cat((o1, o2), dim=1)
-        mask = torch.sigmoid(mask)
-        return modulated_deform_conv(input, offset, mask, self.weight,
-                                     self.bias, self.stride, self.padding,
-                                     self.dilation, self.deformable_groups)
+from ..functions.deform_pool import deform_roi_pooling
 
 
 class DeformRoIPooling(nn.Module):
